+++ changelog.md
@@ -1,1 +1,1 @@
-- Reworked item handling of villager blocks
+- Separate converter and incubator screens